fangall/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
coverage==4.5.1
cryptography==2.3.1
Django==2.1.2
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-redis==4.9.0
djangorestframework==3.9.0
drf-extensions==0.4.0
idna==2.7
kombu==4.2.1
mccabe==0.6.1
Pillow==5.3.0
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
six==1.11.0
sqlparse==0.2.4
vine==1.1.4
